Implemented Bricked Blast Furnace, added corresponding textures
Added a new Casing for the Bricked Blast Furnace, GT_Block_Casings4, ID
15

Refactored the non-electric Blast Furnaces:
The Bronze Blast Furnace and the Bricked Blast Furnace extend a
superclass to ensure consistent behavior.
Subclasses only differ in texture, description, and Casing Block.
