+++ src/main/java/gregtech/common/blocks/GT_Block_Casings4.java
@@ -37,7 +37,8 @@ public class GT_Block_Casings4
         GT_LanguageManager.addStringLocalization(getUnlocalizedName() + ".11.name", "Titanium Turbine Casing");
         GT_LanguageManager.addStringLocalization(getUnlocalizedName() + ".12.name", "Tungstensteel Turbine Casing");
         GT_LanguageManager.addStringLocalization(getUnlocalizedName() + ".13.name", "Engine Intake Casing");
-
+        GT_LanguageManager.addStringLocalization(getUnlocalizedName() + ".15.name", "Dense Bricks");
+        
         ItemList.Casing_RobustTungstenSteel.set(new ItemStack(this, 1, 0));
         ItemList.Casing_CleanStainlessSteel.set(new ItemStack(this, 1, 1));
         ItemList.Casing_StableTitanium.set(new ItemStack(this, 1, 2));
@@ -50,6 +51,7 @@ public class GT_Block_Casings4
         ItemList.Casing_Turbine2.set(new ItemStack(this, 1, 11));
         ItemList.Casing_Turbine3.set(new ItemStack(this, 1, 12));
         ItemList.Casing_EngineIntake.set(new ItemStack(this, 1, 13));
+        ItemList.Casing_DenseBricks.set(new ItemStack(this, 1, 15));
     }
 
     public IIcon getIcon(int aSide, int aMeta) {
@@ -85,7 +87,7 @@ public class GT_Block_Casings4
             case 14:
                 return Textures.BlockIcons.MACHINE_CASING_ROBUST_TUNGSTENSTEEL.getIcon();
             case 15:
-                return Textures.BlockIcons.MACHINE_CASING_ROBUST_TUNGSTENSTEEL.getIcon();
+                return Textures.BlockIcons.MACHINE_CASING_DENSEBRICKS.getIcon();
         }
         return Textures.BlockIcons.MACHINE_CASING_SOLID_STEEL.getIcon();
     }
